# server/internal/api/gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"fmt"
	"time"

	"github.com/gofiber/fiber/v3"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for SessionKind.
const (
	SessionKindTemporary SessionKind = "temporary"
	SessionKindUser      SessionKind = "user"
)

// Valid indicates whether the value is a known member of the SessionKind enum.
func (e SessionKind) Valid() bool {
	switch e {
	case SessionKindTemporary:
		return true
	case SessionKindUser:
		return true
	default:
		return false
	}
}

// Defines values for SessionRole.
const (
	SessionRoleAdmin SessionRole = "admin"
	SessionRoleUser  SessionRole = "user"
)

// Valid indicates whether the value is a known member of the SessionRole enum.
func (e SessionRole) Valid() bool {
	switch e {
	case SessionRoleAdmin:
		return true
	case SessionRoleUser:
		return true
	default:
		return false
	}
}

// Defines values for UpdateUserRequestRole.
const (
	UpdateUserRequestRoleAdmin UpdateUserRequestRole = "admin"
	UpdateUserRequestRoleUser  UpdateUserRequestRole = "user"
)

// Valid indicates whether the value is a known member of the UpdateUserRequestRole enum.
func (e UpdateUserRequestRole) Valid() bool {
	switch e {
	case UpdateUserRequestRoleAdmin:
		return true
	case UpdateUserRequestRoleUser:
		return true
	default:
		return false
	}
}

// Defines values for UserRole.
const (
	UserRoleAdmin UserRole = "admin"
	UserRoleUser  UserRole = "user"
)

// Valid indicates whether the value is a known member of the UserRole enum.
func (e UserRole) Valid() bool {
	switch e {
	case UserRoleAdmin:
		return true
	case UserRoleUser:
		return true
	default:
		return false
	}
}

// AccessCode defines model for AccessCode.
type AccessCode struct {
	Apps        []string           `json:"apps"`
	Code        string             `json:"code"`
	CreatedAt   time.Time          `json:"createdAt"`
	ExpiresAt   time.Time          `json:"expiresAt"`
	Id          openapi_types.UUID `json:"id"`
	Label       string             `json:"label"`
	MaxSignups  int                `json:"maxSignups"`
	RevokedAt   *time.Time         `json:"revokedAt,omitempty"`
	SignupCount int                `json:"signupCount"`
	Url         string             `json:"url"`
}

// AccessCodeList defines model for AccessCodeList.
type AccessCodeList struct {
	Codes []AccessCode `json:"codes"`
}

// AppListRequest defines model for AppListRequest.
type AppListRequest struct {
	Slugs []string `json:"slugs"`
}

// CreateAccessCodeRequest defines model for CreateAccessCodeRequest.
type CreateAccessCodeRequest struct {
	AppSlugs   []string  `json:"appSlugs"`
	ExpiresAt  time.Time `json:"expiresAt"`
	Label      *string   `json:"label,omitempty"`
	MaxSignups int       `json:"maxSignups"`
}

// CreateUserRequest defines model for CreateUserRequest.
type CreateUserRequest struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Fingerprint *string `json:"fingerprint,omitempty"`
	Password    string  `json:"password"`
	Username    string  `json:"username"`
}

// RedeemRequest defines model for RedeemRequest.
type RedeemRequest struct {
	Code        string `json:"code"`
	Fingerprint string `json:"fingerprint"`
	Nickname    string `json:"nickname"`
}

// Session defines model for Session.
type Session struct {
	Apps      []string     `json:"apps"`
	ExpiresAt time.Time    `json:"expiresAt"`
	Kind      SessionKind  `json:"kind"`
	Nickname  *string      `json:"nickname,omitempty"`
	Role      *SessionRole `json:"role,omitempty"`
	Username  *string      `json:"username,omitempty"`
}

// SessionKind defines model for Session.Kind.
type SessionKind string

// SessionRole defines model for Session.Role.
type SessionRole string

// TemporaryAccount defines model for TemporaryAccount.
type TemporaryAccount struct {
	CreatedAt time.Time          `json:"createdAt"`
	CreatedIp string             `json:"createdIp"`
	Id        openapi_types.UUID `json:"id"`
	LastIp    string             `json:"lastIp"`
	Nickname  string             `json:"nickname"`
	RevokedAt *time.Time         `json:"revokedAt,omitempty"`
}

// TemporaryAccountList defines model for TemporaryAccountList.
type TemporaryAccountList struct {
	Accounts []TemporaryAccount `json:"accounts"`
}

// UpdateUserRequest defines model for UpdateUserRequest.
type UpdateUserRequest struct {
	Disabled *bool                  `json:"disabled,omitempty"`
	Password *string                `json:"password,omitempty"`
	Role     *UpdateUserRequestRole `json:"role,omitempty"`
}

// UpdateUserRequestRole defines model for UpdateUserRequest.Role.
type UpdateUserRequestRole string

// User defines model for User.
type User struct {
	Apps     []string           `json:"apps"`
	Disabled bool               `json:"disabled"`
	Id       openapi_types.UUID `json:"id"`
	Role     UserRole           `json:"role"`
	Username string             `json:"username"`
}

// UserRole defines model for User.Role.
type UserRole string

// UserList defines model for UserList.
type UserList struct {
	Users []User `json:"users"`
}

// Id defines model for Id.
type Id = openapi_types.UUID

// CreateAccessCodeJSONRequestBody defines body for CreateAccessCode for application/json ContentType.
type CreateAccessCodeJSONRequestBody = CreateAccessCodeRequest

// RedeemAccessCodeJSONRequestBody defines body for RedeemAccessCode for application/json ContentType.
type RedeemAccessCodeJSONRequestBody = RedeemRequest

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginRequest

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody = CreateUserRequest

// UpdateUserJSONRequestBody defines body for UpdateUser for application/json ContentType.
type UpdateUserJSONRequestBody = UpdateUserRequest

// SetUserAppsJSONRequestBody defines body for SetUserApps for application/json ContentType.
type SetUserAppsJSONRequestBody = AppListRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /api/access-codes)
	ListAccessCodes(c fiber.Ctx) error

	// (POST /api/access-codes)
	CreateAccessCode(c fiber.Ctx) error

	// (POST /api/access-codes/redeem)
	RedeemAccessCode(c fiber.Ctx) error

	// (GET /api/access-codes/{id}/accounts)
	ListTemporaryAccounts(c fiber.Ctx, id Id) error

	// (GET /api/access-codes/{id}/qr)
	GetAccessCodeQR(c fiber.Ctx, id Id) error

	// (POST /api/access-codes/{id}/revoke)
	RevokeAccessCode(c fiber.Ctx, id Id) error

	// (GET /api/access/stream)
	AccessStream(c fiber.Ctx) error

	// (POST /api/login)
	Login(c fiber.Ctx) error

	// (POST /api/logout)
	Logout(c fiber.Ctx) error

	// (GET /api/session)
	GetSession(c fiber.Ctx) error

	// (POST /api/temporary-accounts/{id}/revoke)
	RevokeTemporaryAccount(c fiber.Ctx, id Id) error

	// (GET /api/users)
	ListUsers(c fiber.Ctx) error

	// (POST /api/users)
	CreateUser(c fiber.Ctx) error

	// (PATCH /api/users/{id})
	UpdateUser(c fiber.Ctx, id Id) error

	// (PUT /api/users/{id}/apps)
	SetUserApps(c fiber.Ctx, id Id) error
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []HandlerMiddlewareFunc
}

type MiddlewareFunc fiber.Handler
type HandlerMiddlewareFunc func(c fiber.Ctx, next fiber.Handler) error

// ListAccessCodes operation middleware
func (siw *ServerInterfaceWrapper) ListAccessCodes(c fiber.Ctx) error {

	handler := func(c fiber.Ctx) error {
		return siw.Handler.ListAccessCodes(c)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// CreateAccessCode operation middleware
func (siw *ServerInterfaceWrapper) CreateAccessCode(c fiber.Ctx) error {

	handler := func(c fiber.Ctx) error {
		return siw.Handler.CreateAccessCode(c)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// RedeemAccessCode operation middleware
func (siw *ServerInterfaceWrapper) RedeemAccessCode(c fiber.Ctx) error {

	handler := func(c fiber.Ctx) error {
		return siw.Handler.RedeemAccessCode(c)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// ListTemporaryAccounts operation middleware
func (siw *ServerInterfaceWrapper) ListTemporaryAccounts(c fiber.Ctx) error {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Params("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter id: %w", err).Error())
	}

	handler := func(c fiber.Ctx) error {
		return siw.Handler.ListTemporaryAccounts(c, id)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// GetAccessCodeQR operation middleware
func (siw *ServerInterfaceWrapper) GetAccessCodeQR(c fiber.Ctx) error {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Params("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter id: %w", err).Error())
	}

	handler := func(c fiber.Ctx) error {
		return siw.Handler.GetAccessCodeQR(c, id)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// RevokeAccessCode operation middleware
func (siw *ServerInterfaceWrapper) RevokeAccessCode(c fiber.Ctx) error {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Params("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter id: %w", err).Error())
	}

	handler := func(c fiber.Ctx) error {
		return siw.Handler.RevokeAccessCode(c, id)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// AccessStream operation middleware
func (siw *ServerInterfaceWrapper) AccessStream(c fiber.Ctx) error {

	handler := func(c fiber.Ctx) error {
		return siw.Handler.AccessStream(c)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// Login operation middleware
func (siw *ServerInterfaceWrapper) Login(c fiber.Ctx) error {

	handler := func(c fiber.Ctx) error {
		return siw.Handler.Login(c)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// Logout operation middleware
func (siw *ServerInterfaceWrapper) Logout(c fiber.Ctx) error {

	handler := func(c fiber.Ctx) error {
		return siw.Handler.Logout(c)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// GetSession operation middleware
func (siw *ServerInterfaceWrapper) GetSession(c fiber.Ctx) error {

	handler := func(c fiber.Ctx) error {
		return siw.Handler.GetSession(c)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// RevokeTemporaryAccount operation middleware
func (siw *ServerInterfaceWrapper) RevokeTemporaryAccount(c fiber.Ctx) error {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Params("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter id: %w", err).Error())
	}

	handler := func(c fiber.Ctx) error {
		return siw.Handler.RevokeTemporaryAccount(c, id)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// ListUsers operation middleware
func (siw *ServerInterfaceWrapper) ListUsers(c fiber.Ctx) error {

	handler := func(c fiber.Ctx) error {
		return siw.Handler.ListUsers(c)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// CreateUser operation middleware
func (siw *ServerInterfaceWrapper) CreateUser(c fiber.Ctx) error {

	handler := func(c fiber.Ctx) error {
		return siw.Handler.CreateUser(c)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// UpdateUser operation middleware
func (siw *ServerInterfaceWrapper) UpdateUser(c fiber.Ctx) error {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Params("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter id: %w", err).Error())
	}

	handler := func(c fiber.Ctx) error {
		return siw.Handler.UpdateUser(c, id)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// SetUserApps operation middleware
func (siw *ServerInterfaceWrapper) SetUserApps(c fiber.Ctx) error {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Params("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter id: %w", err).Error())
	}

	handler := func(c fiber.Ctx) error {
		return siw.Handler.SetUserApps(c, id)
	}

	for i := len(siw.HandlerMiddlewares) - 1; i >= 0; i-- {
		m := siw.HandlerMiddlewares[i]
		next := handler
		handler = func(c fiber.Ctx) error {
			return m(c, next)
		}
	}

	return handler(c)
}

// FiberServerOptions provides options for the Fiber server.
type FiberServerOptions struct {
	BaseURL            string
	Middlewares        []MiddlewareFunc
	HandlerMiddlewares []HandlerMiddlewareFunc
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router fiber.Router, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, FiberServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router fiber.Router, si ServerInterface, options FiberServerOptions) {
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.HandlerMiddlewares,
	}

	for _, m := range options.Middlewares {
		router.Use(fiber.Handler(m))
	}

	router.Post(options.BaseURL+"/api/login", wrapper.Login)

	router.Post(options.BaseURL+"/api/logout", wrapper.Logout)

	router.Get(options.BaseURL+"/api/session", wrapper.GetSession)

	router.Post(options.BaseURL+"/api/access-codes/redeem", wrapper.RedeemAccessCode)

	router.Get(options.BaseURL+"/api/users", wrapper.ListUsers)

	router.Post(options.BaseURL+"/api/users", wrapper.CreateUser)

	router.Patch(options.BaseURL+"/api/users/:id", wrapper.UpdateUser)

	router.Put(options.BaseURL+"/api/users/:id/apps", wrapper.SetUserApps)

	router.Get(options.BaseURL+"/api/access-codes", wrapper.ListAccessCodes)

	router.Post(options.BaseURL+"/api/access-codes", wrapper.CreateAccessCode)

	router.Post(options.BaseURL+"/api/access-codes/:id/revoke", wrapper.RevokeAccessCode)

	router.Get(options.BaseURL+"/api/access-codes/:id/qr", wrapper.GetAccessCodeQR)

	router.Get(options.BaseURL+"/api/access-codes/:id/accounts", wrapper.ListTemporaryAccounts)

	router.Post(options.BaseURL+"/api/temporary-accounts/:id/revoke", wrapper.RevokeTemporaryAccount)

	router.Get(options.BaseURL+"/api/access/stream", wrapper.AccessStream)

}